price history
implemented the get_trades query in queries.py
basic implementation of getting the trades history of a given asset
(currently only in usd). Tests with ETH, which takes a little while to
load,
read/write history functions use a csv with the id of the last trade
loaded as the first entry, and follows with a list of trades.
Need to clean up get_history to work with any valid pair and to have it
use the read/write history methods to only update as necessary